historic data is collected in different container now. docker-compose.yml is modified according to this change, setup.sh also. Renaming module in order to be more mean
